2022 10 24 interpreter interface

+++ contracts/interpreter/ops/core/OpContext.sol
@@ -2,9 +2,9 @@
 pragma solidity ^0.8.15;
 
 import {IERC20Upgradeable as IERC20} from "@openzeppelin/contracts-upgradeable/token/ERC20/IERC20Upgradeable.sol";
-import "../../runtime/LibStackTop.sol";
-import "../../runtime/LibInterpreterState.sol";
-import "../../integrity/LibIntegrityState.sol";
+import "../../run/LibStackTop.sol";
+import "../../run/LibInterpreterState.sol";
+import "../../deploy/LibIntegrityState.sol";
 import "../../../idempotent/LibIdempotentFlag.sol";
 
 /// @title OpContext
@@ -24,12 +24,14 @@ library OpContext {
         StackTop stackTop_
     ) internal pure returns (StackTop) {
         uint256 column_ = Operand.unwrap(operand_) >> 8;
-        uint256 row_ = Operand.unwrap(operand_) & uint(type(uint8).max);
-        integrityState_.contextScratch = IdempotentFlag.unwrap(LibIdempotentFlag.set16x16(
-            IdempotentFlag.wrap(integrityState_.contextScratch),
-            column_,
-            row_
-        ));
+        uint256 row_ = Operand.unwrap(operand_) & uint256(type(uint8).max);
+        integrityState_.contextScratch = IdempotentFlag.unwrap(
+            LibIdempotentFlag.set16x16(
+                IdempotentFlag.wrap(integrityState_.contextScratch),
+                column_,
+                row_
+            )
+        );
         // Note that a expression with context can error at runtime due to OOB
         // reads that we don't know about here.
         return integrityState_.push(stackTop_);
